requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-jose
requests
pydantic-settings
psycopg2